Rename moveable folder to movable for consistency.
